Fixing http/https issue with Streamus
When Streamus was first created there was no such thing as a YouTube
iFrame API. As such, it was created in a way that mimics the iFrame API
while not actually going through it. This resulted in some hard-to-spot
inconsistencies surrounding http/https. In very niche scenarios, a user
would be unable to use Streamus if their configuration was such that
they could only use http or https. By going through the iFrame API we
can ensure that the correct tools are loaded for a given client.