Bumping Chrome minimum version, re-adding 'Reload', patching YouTube API
YouTube's API broke and all songs were saying "Unable to be played"
unless I load the player with a song by default. So, I grabbed a random
song and the player now loads it. Shouldn't be apparent to the end users
at all.

Also, chrome.extension.reload no longer crashes the browser in Chrome
v40, so I re-added it.